+++ src/main/java/ru/practicum/shareit/item/ItemRepository.java
@@ -6,13 +6,13 @@ import ru.practicum.shareit.item.dto.ItemUpdateDto;
 import java.util.*;
 
 public interface ItemRepository {
-    ItemDto addItem(ItemDto itemDtoRequest, Integer userId);
+    ItemDto addItem(ItemDto itemDtoRequest, Long userId);
 
-    ItemDto updateItem(Integer itemId, ItemUpdateDto itemDtoRequest, Integer userId);
+    ItemDto updateItem(Long itemId, ItemUpdateDto itemDtoRequest, Long userId);
 
-    ItemDto getItem(Integer itemId);
+    ItemDto getItem(Long itemId);
 
-    List<ItemDto> getOwnerItems(Integer ownerId);
+    List<ItemDto> getOwnerItems(Long ownerId);
 
     List<ItemDto> itemSearch(String text);
 }
