+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -3,8 +3,6 @@ package ru.practicum.shareit.item.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.shareit.booking.Booking;
-import ru.practicum.shareit.booking.StatusBooking;
 import ru.practicum.shareit.request.ItemRequest;
 import ru.practicum.shareit.user.model.User;
 
@@ -15,7 +13,7 @@ import ru.practicum.shareit.user.model.User;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Item {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private Boolean available;
